+++ Sources/MapboxMaps/Gestures/GestureHandlers/PinchBehaviors/PanZoomPinchBehavior.swift
@@ -13,14 +13,13 @@ internal final class PanZoomPinchBehavior: PinchBehavior {
         self.mapboxMap = mapboxMap
     }
 
-    internal func update(pinchMidpoint: CGPoint,
-                         pinchScale: CGFloat,
-                         pinchAngle: CGFloat) {
+    internal func update(pinchMidpoint: CGPoint, pinchScale: CGFloat) {
         mapboxMap.performWithoutNotifying {
             mapboxMap.setCamera(
                 to: CameraOptions(
                     center: initialCameraState.center,
-                    zoom: initialCameraState.zoom))
+                    zoom: initialCameraState.zoom,
+                    bearing: initialCameraState.bearing))
 
             mapboxMap.dragStart(for: initialPinchMidpoint)
             let dragOptions = mapboxMap.dragCameraOptions(